Reorganize to folders
This will make it easier to do styles on a per-component basis
