+++ src/App.js
@@ -1,8 +1,10 @@
 import React from 'react'
-import Main from './Main'
+import Main from './main'
+import Header from './header'
 
 const App = () => (
 	<div>
+		<Header />
 		<Main />
 	</div>
 )
