+++ src/header/index.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+class Header extends React.Component {
+	render() {
+		return (
+			<div>
+				<h2> Home </h2>
+				<h2> Live </h2>
+				<h2> Admin </h2>
+			</div>
+		)
+	}
+}
+
+export default Header
